Tweak nix and stack set up
At the moment IHaskell can't be built with nixos-18.09 because of changes
to the zeromq library which causes the Haskell bindings to fail. This
has been "fixed" in the zermoq4-haskell bindings but it's not yet been
released, so the nix setup grabs a version from GitLab. Unfortunately
I then have communication problems when using ihaskell-notebook from the
nix shell, so I don't know if the "fix" causes problems or there is
something else amiss.

Bumped the stack LTS to ghc 8.4.4 (LTS 12.26).