Add IHaskell.Display.HvegaLab for Jupyter Lab
Since Jupyter Lab has native support for Vega and Vega-Lite visualizations,
the display code is much simpler. I felt it was best to create a separate
module as it's not yet clear how much the IHaskell community has moved
from "classic" (notebook) to "new" (lab).

This change requires the latest version of IHaskell (0.9.1.0) to get
at the improved mimetype support for IHaskell.Display.

TODO: add a PNG option (if possible) to make it possible to see the
visualizations ia a notebook viewer.