Extend ProjectionProperty
The PPrecision value is apparently encoded as a string, not a number,
in the 3.3.0 Vega Lite specification, so this has been fixed in this
commit.

Note that all the "custom" projection tests create invalid VegaLite
specifications, as the type list is fixed in the spec, with no
(obvious to me at least) way to say "hey, this is a custom job".