Breaking change: change MBorders from taking a bool to [MarkProperty]
The MBorders constructor was added in #10, where it took a Bool.
Elm has it take a list of MarkProperty, which is what this commit
does.

Looking at version 3.3.0 of the Vega-Lite schema, I see

        "borders": {
          "anyOf": [
            {
              "type": "boolean"
            },
            {
              "$ref": "#/definitions/MarkConfig"
            }
          ]
        }

which means both are supported, but I can't find any documentation
on what a boolean means, so for now I have dropped the boolean
support (this also holds for a number of other constructors that
take [MarkProperty]).