ScaleProperty: SScheme improvements
The SScheme constructor was labelled as supporting a single argument
in the optional list, but this was not handled correctly (setting the
count argument).

Support for setting both count and extent, by supplying three numbers,
has been added.

This should probably be re-worked to add a data type to clearly reflect
the #/definitions/SchemeParams object rather than this ad-hoc solution.