Add tests based on elm vega-lite
These are regression tests based on some of the test code from the
ElmVega Lite package.

The travis-CI testing is based on

    https://chshersh.github.io/posts/2019-02-25-haskell-travis

with slight tweaks because of the multi-project structure of this
repository (although only one repository has tests).

Many of the tests are "incomplete", since this version of hvega is
missing necessary functionality. This can mean missing tests, or
"partial" tests. Look for "TODO" comments in the tests. Several tests
have not yet been added, because they would test nothing at this time
(see comments in tests/Tests.hs).

The ConfigTests currently do not set padding (unlike the Elm version)
because this generates an invalid VegaLite specification.

A number of other tests generate invalid VegaLite specifications, but
they have been left as is:

- fillsroke/fill1 /stroke1 /combined1

  create empty objects for the properties

- color/interp1 to color/interp5

  fail as using invalid SInterpolate arguments

- color/scale2

  uses the deprecated ScSequential constructor