Special case: named styles for a singleton list and Count operation
This should have no semantic change to the output, since the Vega-Lite
specification treats "[string]" and "string" as the same for the
"styles" field, and allows "field" to be unspecified with the "count"
option.

It has been made to allow easier comparison of the hvega output to the
Vega-Lite example gallery.