Documentation tweaks
Mainly adding links to examples, for symbols that are either being
defined or have already been used in the example. This is to catch
the case when an invalid symbol has been used (as has been done
before) and may be less confusing to readers (i.e. "why is X a link
in this line but not in the next one?"). This is not complete.

Added some extra documentation to link together dataRow/Column routines
and types.