Add a ZIndex type
This replaces the ZIndex new type introduced in the 0.4.0.0 development
series with an actual type. The main use is ZBack and ZFront, but ZValue
is added as a fall bak, in case the specification is expaned (or systems
start making use of zindex values that are not 0 or 1).

Added a test that uses ZIndex, as there weren't any previously.