Add the Angle type alias
It is not clear to me if all angles (that have been labelled with this
type) are measured from the same location.

There are several other angle values in the API that have not been
labelled, since they are in radians, or are not guaranteed to match
the use case of this type. This is worth looking at at some point.