Add the Gaia dataset to the source directory
This has been added so that it is available for the tutorial. Since
the tutorial uses http access (at least at present) the file has
not been added to the .cabal file as either a requirement for a source
build or a data file for the installation. This may change.