Add a simple validation test
It's useful to check whether the test output is actually valid, and
this is annoying to do manually (particularly because, for reasons,
some of the tests don't generate valid files). This script is a
simple wrapper around the Python jsonschema package to validate the
test files, and skip "known failures". Where the known failures
are against version 3.4.0 of the Vega-Lite specification.