Breaking change: replace RemoveInvalid by MRemoveInvalid
In Vega-Lite version 4, the handling of "invalid" values has changed
from being a top-level configuration option to being part of the
mark category, so RemoveInvalid has been removed from
ConfigurationProperty and replaced with MRemoveInvalid in
MarkProperty.

All of the tests of this functionality had set the flag to False,
so this commit also changes one of them to use True (to improve
coverage). The changes to the test specifications have been
validated against version 4.0.2 of the Vega-Lite schema.