Haddock: list routines that can be used with encoding
This started out as trying to add the "key" encoding channel, based
on Elm, but I am confused about the type (and use), so this has been
shelved. There are some minor documentation improvements instead
(and commented-out code for the key channel, in case I want to add
it soon).