BREAKING CHANGE: lookup has changed its type, lookupAs is deprecated
The lookup encoding has been changed so that the list of fields
stored when the keys match is now specified by the LookupFields
type (rather than a list of field names). This supports providing
aliases and handling of default values, and now subsumes the lookupAs
encoding, which has been marked as deprecated.