Simplify use of RgExtent and DnExtent
Rather than take two DataValue arguments, take two Doubles, as this
is all that is supported by Vega Lite. I had explored using
existential quantification to support separating the extent types
from the conversion to a DataValue (to make it clear that the two
values should have the same underlying type), then realised that
this was excessive.