Replace ConcatConfig and FacetConfig by CompositionConfig
These types were used by ConcatStyle and FacetStyle, but - circa v4.4.0 -
there's no reason to have them as separate types as they are essentially
identical. This type will also be used to add support for RepeatStyle
in a future commit.

This change also helps with an earlier change in the 0.6.0.0 release,
where the addition of FSpacing to FacetChannel caused the FacetStyle
constructors to be renamed.