Add more MNoXXX variants for MarkProperty
Error box and band properties (constructors in MarkProperty) can now
be turned off with explicit "No" variants: MNoBorders, MNoBox,
MNoMedian, MNoRule, and MNoTicks. These join the MNoOutliers
constructor.

The "set the property" variants of these, that take a list, will now
create "true" rather than "{}" for the output when given the empty
list. This should be semantics preserving (as far as I understand
the schema).