AxisConfig: add Disable (Vega-Lite 4.8)
Add the Disable constructor to AxisConfig.

In testing I couldn't seem to get this setting to do anything
(i.e. I had to keep the 'PAxis []' settings). Have I mis-understood
things or has the JS plotting code not been updated to support this
yet? I have left the tests with calls to both Disable and PAxis for
now.