BREAKING CHANGE: rework time-unit handling
The TimeUnit and ScaleNice types have now been split into two
parts: a type that has the "basic" times (e.g. milli-second), which
is then combined with optional types in the "top-level" type. This
means that

    PTimeUnit Year
    PTimeUnit (Utc DayHours)
    SNice NMinute

are now

    PTimeUnit [TU Year]
    PTimeUnit [Utc DayHours]
    SNice (NTU NMinute)

This is more verbose, but avoids some problems with the old
system (like infinite recursion in the 'Utc xxx' constructor).