TimeUnit: switch back to a scalar rather than list argument
I believe that the timeUnit field can be represented as a "scalar",
that is, all input values specified in a single constructor, since
the TUMaxBins argument is (if I'm reading this correctly) only
for use when no other argument is given. This is hopefully a little
bit easier to use.