Switch to nix setup from hs-nix-template
See the cookiecutter directory in
  https://github.com/utdemir/hs-nix-template

Also updated the niv version. This is currently only used in the
top-level shell.nix which unfortunately does not work as ihaskell
is marked as broken. The hvega and ihaskell-hvega packages use the
unstable package setup, as that has haskell-language-server in
it (but the ihaskell-hvega setup is also broken thanks to
ihaskell in nix).