Tests: support hashable 0.3.1.0
The test code noted that the tests would fail if the conversion
of an Aeson value to "pretty" text changed, but I hadn't noted
that the defauld "pretty" encoding was susceptible to changes
in the underlying code. And guess what hashable 0.3.1.0 did?

The tests should now hopefully be independent of changes to
the internals of Aeson. Fingers crossed.